Task 3 - Aspect Ratio
Our cards are currently arranged in a single row (we’ll fix that next week).  That’s making our cards really tall and skinny (especially in portrait) which doesn’t look very good.  Force each card to have a width to height ratio of  2:3 (this will result in empty space above and/or below your cards, which is fine).