Task 5 - Use smaller font when 5 pairs
When your game randomly shows 5 pairs, the font we are using for the emoji will be too large (in portrait) and will start to get clipped.  Have the font adjust in the 5 pair case (only) to use a smaller font than .largeTitle.  Continue to use .largeTitlewhen there are 4 or fewer pairs in the game.