Create Deal 3 More Cards button
- You will need to have a “Deal 3 More Cards” button (per the rules of  Set).
a. when it is touched, replace the selected cards if  the selected cards make a Set (with fly-in/fly-away as described above)
b. or, if  the selected cards do not make a Set (or if  there are fewer than 3 cards selected, including none), fly in (i.e. animate the arrival of) 3 new cards to join the ones already on screen (and do not affect the selection)
c. disable this button if  the deck is empty